src/requirements.txt
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.1
torch>=1.7.0  # see https://pytorch.org/get-started/locally (recommended)
torchvision>=0.8.1